requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
Django==3.2.4
djangorestframework==3.12.4
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2